Re #1519: Misc fixes in pjsua2:  - Changed AccountConfig::toPj() signature, because a pjsua_acc_config instance should not be copied/assigned to another pjsua_acc_config instance as it contains list.  - Fixed assertion in Endpoint destructor caused by access to pjlib mutex after pjlib is destroyed.  - Changed VS2005 sample_debug project to allow debugging C++ code such as pjsua2_demo.cpp.  - Compile warnings on VS2005.  - Fixed SWIG python unit test.
git-svn-id: https://svn.pjsip.org/repos/pjproject/branches/projects/pjsua2@4692 74dad513-b988-da41-8d7b-12977e46ad98

+++ pjlib/src/pj/string.c
@@ -189,7 +189,7 @@ PJ_DEF(float) pj_strtof(const pj_str_t *str)
     part.slen = pdot ? pdot - str->ptr : str->slen;
 
     if (part.slen)
-	val = pj_strtol(&part);
+	val = (float)pj_strtol(&part);
     else
 	val = 0;
 
@@ -199,8 +199,8 @@ PJ_DEF(float) pj_strtof(const pj_str_t *str)
 	if (part.slen) {
 	    pj_str_t endptr;
 	    float fpart, fdiv;
-	    unsigned i;
-	    fpart = pj_strtoul2(&part, &endptr, 10);
+	    int i;
+	    fpart = (float)pj_strtoul2(&part, &endptr, 10);
 	    fdiv = 1.0;
 	    for (i=0; i<(part.slen - endptr.slen); ++i)
 		    fdiv = fdiv * 10;
